// src/types/structures/guild.ts
export interface GuildMember {
  /** The user this guild member represents */
  user?: User;
  /** This user's guild nickname */
  nick?: string | null;
  /** The member's guild avatar hash */
  avatar?: string | null;
  /** Array of role object IDs */
  roles: string[];
  /** When the user joined the guild */
  joined_at: string;
  /** When the user started boosting the guild */
  premium_since?: string | null;
  /** Whether the user is deafened in voice channels */
  deaf: boolean;
  /** Whether the user is muted in voice channels */
  mute: boolean;
  /** Guild member flags represented as a bit set */
  flags: number;
  /** Whether the user has not yet passed the guild's Membership Screening requirements */
  pending?: boolean;
  /** Total permissions of the member in the channel, including overwrites, returned when in the interaction object */
  permissions?: string;
  /** When the user's timeout will expire and the user will be able to communicate in the guild again */
  communication_disabled_until?: string | null;
}